chore(webPreload): Remove Duplicate code from WebViewPreload
